Remove unused GetDexRegisterMap arguments.
They are no longer needed in the new encoding.

I reuse the local variables in most places to DCHECK the size
of the decoded register map.  This has one catch though:
We sometimes omit all dex registers, so the DCHECK should be
done only after checking if the map is empty (if applicable).

Test: test-art-host-gtest-stack_map_test
Change-Id: I94b67029842374bc8eb7c9e5eac76fc93a651f24