Remove useless assertions in ConcurrentCopying::MarkNonMoving.
These assertions were checking that the read barrier state was
"marked" ("gray" or "black") when the reference passed as argument was
already marked in the corresponding bitmap. However, when the read
barrier "black" state was folded into the "white" state, these
assertions essentially became checks verifiying that the read barrier
state was "gray" or "non-gray", which is a tautology.

Test: art/test.py
Bug: 110812755
Change-Id: Icec68c85de88b3f188b23a2d59bdd5dd60cc72ed