Update test expectations for 911-get-stack-trace.
D8 version 1.3.8-dev has become less aggressive in applying
redundant field load elimination. Therefore, there is a load
that is no longer eliminated which changes the offsets.

When we added the more aggressive field load elimination
we did the reverse change here:

https://android-review.googlesource.com/c/platform/art/+/707549

This is going back to what it was before.

Test: art/test/testrunner/testrunner.py -b --host -t 911-get-stack-trace
Change-Id: I80194d605ce85705562e7d3e2642d04ec90575dc