Revert^4 "Rewrite JVMTI method tracing to use trampolines"
Since we are unable to change the semantics of ClassLoaded and
ClassPrepared due to b/111394423 for the moment changes were made in
libjdwp to remove the deadlock between the jit being stopped and
jit-threads trying to lock JVMTI monitors in libjdwp.

Until some resolution for b/111394423 is implemented agents should be
very careful about gaining locks in the ClassLoaded, ClassPrepared or
VMObjectAlloc events if said lock is held over a call to
SetEventNotificationMode, SetExtensionEventCallback,
RetransformClasses, or RedefineClasses. These locks can only be
safely held if the object argument to the event is not an array-class
object or the thread is known not to be a jit thread.

This reverts commit bf2ca05ccf658a9f626d1cb3f1adca6087ece7b8.

Bug: 111394423
Test: ./art/tools/run-libjdwp-tests.sh --mode=host --debug

Reason for revert: Fixed issue with libjdwp deadlocking on jit threads

Change-Id: I8baeb8311f67062cbc26286eb221da34e1972f27