Don't try to make /system on device writable when syncing ART to chroot dir.
When using a chroot dir for ART testing, we are only pushing files to
directory $ART_TEST_CHROOT (which is expected to be under /data on the
device) -- /system is left untouched, which means we do not need to
try and make it writable before syncing ART to the device for testing
purposes.

Note that skipping this step is especially useful in the case of
chroot-based testing, as making /system writable requires a reboot on
most devices nowadays (it's needed after running
`adb disable-verity`), whichs resets some of the chroot setup (and
thus requires running the chroot setup script again).

Test: Rely on the ART Buildbot
Bug: 34729697
Change-Id: Id4e4ad93a888a1180c0502e541b9497cfa58d211