Add specialized handler for non-quickened opcodes in mterp.
The generic handler did access checks but this is unnecessary since
mterp should only be used in cases where we don't need access check.
I believe that this is just simple oversight that was never spotted.

Non-quickened code is currently 1.9x slower than quickened code.
This reduces it to 1.45x - essentially halving of the overhead.
(golem benchmarks on arm64 with quickening manually disabled)

Test: test-art-host-gtest
Change-Id: Icd29aeff3bf25baaae9a9a07072ddd3448e4d92c