ART: Improve test 160-read-barrier-stress.
Add tests for reading references for which we do not hold a
to-space reference. Rewrite the test to test fields, arrays,
GC roots and Unsafe separately, so that reads in one of the
tests do not affect other tests and we exercise more code
paths in the Baker read barriers with introspection.

Add a regression test for the reverted
    https://android-review.googlesource.com/729224
which erroneously clobbered r0 on ARM in UnsafeCASObject.

Test: testrunner.py --target 160-read-barrier-stress
Test: Repeat with the bad CL; fails for ARM, OK for ARM64.
Bug: 36141117
Change-Id: I0cbc5f5b9a78865ba7ca14ae166ee19d7cf8dd3f